factors the list header out into requirement details and adds styling
