+++ src/components/pages/RequirementDetails/index.scss
@@ -1,16 +1,86 @@
 @import "../../../styles/variables";
 
 .requirement-detail-container {
-  display: flex;
-  justify-content: center;
-  position: relative;
-  box-sizing: border-box;
-  @media screen and(max-width: $breakpoint-medium) {
-    flex-direction: column;
-    align-items: center;
-    left: 0;
+  width: 100%;
+  .resources-steps {
+    display: flex;
+    justify-content: center;
+    position: relative;
+    box-sizing: border-box;
+    @media screen and(max-width: $breakpoint-medium) {
+      flex-direction: column;
+      align-items: center;
+      left: 0;
+    }
+    @media screen and(max-width: $breakpoint-medium) {
+      width: 100%;
+    }
   }
-  @media screen and(max-width: $breakpoint-medium) {
+
+  .list-header {
+    color: white;
     width: 100%;
+    text-align: center;
+    @media screen and (min-width: $breakpoint-medium) {
+      display: grid;
+      grid-template-columns: 1fr 1fr 1fr 1fr 1fr;
+    }
+    @media screen and (min-width: $breakpoint-large) {
+    }
+    .step-description {
+      @media screen and (min-width: $breakpoint-medium) {
+        align-self: center;
+        grid-column: 2 / span 3;
+      }
+      p {
+        padding-top: 1rem;
+      }
+    }
+    .progress-ring {
+      justify-self: center;
+      margin: auto;
+      width: 100%;
+      box-sizing: border-box;
+      padding: 3rem;
+      font-size: 3rem;
+      max-width: 150px;
+      @media screen and (min-width: $breakpoint-medium) {
+        grid-column: 5 / span 1;
+        padding: 2rem;
+        // font-size: 2rem;
+      }
+      @media screen and (min-width: $breakpoint-large) {
+        padding: 3rem;
+        max-width: 200px;
+      }
+      .progress-number,
+      .progress-label {
+        position: relative;
+        top: -0.2rem;
+        @media screen and (min-width: $breakpoint-medium) {
+        }
+        @media screen and (min-width: $breakpoint-large) {
+          top: 0;
+          // top: -0.3rem;
+        }
+      }
+    }
+  }
+
+  // .icon-title {
+  //   display: flex;
+  //   align-items: center;
+  //   @media screen and(max-width: $breakpoint-small) {
+  //     width: 100%;
+  //   }
+  // }
+
+  .title {
+    // background: none;
+    font-size: 3.2rem;
+    font-weight: 800;
+    @media screen and(max-width: $breakpoint-small) {
+      padding-top: 0;
+    }
   }
 }
